Add rad_scoped_enum_helpers
Various helpers for scoped enums (aka enum classes). Currently this header provides the RAD_ENABLE_SCOPED_ENUM_BITWISE_OPS macro, which expands to bitwise operators for the given scoped enum type. This allows you to use a scoped enum with all bitwise operations, such as bitwise AND ( & ) or bitwise OR ( | ). The result of bitwise AND is even explicitly cast-able to bool, so you can do, for example: `if (flags & foo_flags::some_option)`.